+++ nodes/EasyRedmine/client/Project.ts
@@ -1,0 +1,11 @@
+export interface Project {
+	id: number;
+	name: string;
+	status: number;
+	trackers: EasyProjectTracker[] | undefined;
+}
+
+export interface EasyProjectTracker {
+	id: number;
+	name: string;
+}